create an IndividualsTableView component
